requirements.txt
PyGithub == 1.57
python-sonarqube-api == 1.3.5